Reply no longer flips fields in its input data to generate a reply. Instead, it creates an object in the obvious manner, from its arguments. The previous "clever" behaviour is now available as a function, reply_to(). This avoids making the class *too* clever (and too confusing).
If a Message (or subclass) is created from another message (or data array, or
string, or tuple) then any specific arguments (for 'to', etc.) will override
the values specified in the "bulk" argument. This should make message
construction easier to understand.

Note that reading a message from a KSock still returns (just) a Message -
returning an appropriate subclass would be nice, but always involves some extra
cost, and it may not be something the caller wants. Thus, for the moment, the
user is left to call the 'cast()' method to achieve this result, if they wish.

--HG--
extra : convert_revision : svn%3A53f7ac7a-121e-11de-a074-998af0e899fe/trunk%4098