A new NewMessage class, and supporting infrastructure.
This is a first stab at something that may replace the current
Message class. For the moment, it does not do so, and there's
no proof yet that it can do so. As committed, the new code is
tested as itself, but not as a replacement for Message, nor
is it actually *used* for anything.

Only checkout if you really want to - it shouldn't make any
material difference over the previous checking.

--HG--
extra : convert_revision : svn%3A53f7ac7a-121e-11de-a074-998af0e899fe/trunk%40116