Various fixes.
1. Correct order of events at KBUS exit

Now I've moved to using the proper kernel debugging message
functions, it becomes clear (because the device names came
out wrong in dmesg) that I had the order of stuff in
_exit_kbus_exit() at least a bit wrong. Fix it.

2. Remove redundant device index from messages

It's not needed because dev_dbg and friends output the device
name as part of their own prefixing, and KBUS names the devices
with their index.

3. In kbus_report.c, add in copyright header, because I feel I should.

4. Keep a copy of the Kconfig here for later use

Indent the Kconfig help text by two spaces
as is mandated by the documentation on Kconfig files.