Amend documentation so Sphinx 1.1.2 doesn't crash
ReadTheDocs are running Sphinx 1.1.2. This crashes if one tries
to use :const:SOMETHING where it can't find the SOMETHING - for
instance, in some of the functions in messages.py, which are
referring to constants defined in the Message class.

Fix the problem by amending the docstrings to be more explicit.