WaitForMessage() now exits on ksock closure.
Added code to cause WaitForMessage() to exit with -EINTR when the ksock
it's waiting on is closed from another thread (Linux does not provide
this behaviour by default - waiting on a closed file descriptor just
waits, unless or until the descriptor gets reused).

The implementation uses an eventfd which is written to in the close()
call, causing any poll() in WaitForMessage() to exit.